refactor from loop to list comprehension: type (pauses)
This is a pure refactor, maintaining the existing behavior, but it looks
like there might be a bug in how the Waypoint.Type is assigned when
there is a pause. I would expect the type to be WaypointType.Pause
during the whole paused segment, but Pause is only set for the beginning
of the paused segment, then the rest of the paused segment uses
WaypointType.Regular until the paused segment ends, at which point we
get a single entry of WaypointType.Resume.

There is also some weird behavior when the activity starts in a paused
state.